Add files via upload
# Implementing Batch Gradient Descent
The following .ipynb file contains my own implementation of the working of Batch Gradient Descent for Linear Regression run on the dataset provided by ScikitLearn (California Housing). This is a "from scratch" implementation of the algorithm without the use of any external libraries or environments like PyTorch or TensorFlow. 
# Why a from scratch demonstration?
I believe practical understanding and visualization to be an essential part of learning. This demonstration allows beginners and experimenters to experience the process, aka how things really work under the hood. 
# What exactly did I do?
Batch Gradient Descent is an algorithm to estimate the parameters required for making a "predicting function" for a dataset with given (input, target) pairs. It uses the concept of hyperdimensional space and explores the effect of learning rates and number of epochs/iterations on the overall parametric convergence. For my own personal experience, refer to the my_experience.md file!
# My own observations:
1. The convergence plot of theta0 and theta1 showed a straight and stable line with a small upwards hook towards the end. This shows a stable trajectory towards convergence, as expected.
2. The error started to increase between actual and predicted values as actual values increased.

Well that's pretty much it. Thank you for sticking around!